Ejemplos

+++ ejemplos-while/EjemploCicloWhile/src/ejemplociclowhile/Ejemplo032.java
@@ -23,7 +23,7 @@ public class Ejemplo032 {
         entrada.useLocale(Locale.US);
         int limite = 5;
         int contador = 1;
-        int suma = 0;
+        double suma = 0;
         int valor_ingresado;
         double promedio;
         
@@ -35,7 +35,7 @@ public class Ejemplo032 {
             contador = contador + 1;
         }
         promedio = suma / limite;
-        System.out.printf("La suma final es %d\n", suma);
+        System.out.printf("La suma final es %.2f\n", suma);
         System.out.printf("EL promedio final es %.2f\n", promedio);
 
     }
